Expand gh action tests (#259)
* expand testing

* add windows testing support

* fix tests

* fix windows tests

* fix data path

* fix lint

* change data paths and column reordering

* force int64

* remove oriondbexplorer tutorial from test

* change kernel and remove dtype check

* add matplotlib to dep

* update tox docs command

+++ orion/core.py
@@ -97,8 +97,8 @@ class Orion:
     @staticmethod
     def _build_events_df(events):
         events = pd.DataFrame(list(events), columns=['start', 'end', 'severity'])
-        events['start'] = events['start'].astype(int)
-        events['end'] = events['end'].astype(int)
+        events['start'] = events['start'].astype('int64')
+        events['end'] = events['end'].astype('int64')
 
         return events
